RHOAIENG-65096: Lift connection type types and utils to k8s-core (Phase 1)
Move all connection type type definitions, enums, and stateless utility
functions from frontend/src/concepts/connectionTypes/ to
@odh-dashboard/k8s-core. Replace originals with backward-compatible
re-export shims. Update all package-level consumers to import directly
from k8s-core.

Phase 2 will add extension points for data-fetching hooks and move
shared UI components.
