+++ packages/cypress/cypress/utils/modelServingUtils.ts
@@ -21,7 +21,7 @@ import {
   mockServingRuntimeTemplateK8sResource,
   mockInvalidTemplateK8sResource,
 } from '@odh-dashboard/internal/__mocks__/mockServingRuntimeTemplateK8sResource';
-import { ConnectionTypeFieldType } from '@odh-dashboard/internal/concepts/connectionTypes/types';
+import { ConnectionTypeFieldType } from '@odh-dashboard/k8s-core';
 import { ServingRuntimePlatform } from '@odh-dashboard/model-serving/shared/types';
 import { DataScienceStackComponent } from '@odh-dashboard/plugin-core/areas';
 import { mockServiceAccountK8sResource } from '@odh-dashboard/internal/__mocks__/mockServiceAccountK8sResource';
