+++ packages/k8s-core/src/index.ts
@@ -107,3 +107,68 @@ export {
   getGeneratedSecretName,
   isGeneratedSecretName,
 } from './secretUtils';
+
+export {
+  ConnectionTypeFieldType,
+  connectionTypeDataFields,
+  AwsKeys,
+  AccessTypes,
+} from './connectionTypes';
+export type {
+  ConnectionTypeDataFieldTypeUnion,
+  ConnectionTypeFieldTypeUnion,
+  ConnectionTypeCommonProperties,
+  DataField,
+  SectionField,
+  HiddenField,
+  ShortTextField,
+  TextField,
+  UriField,
+  FileField,
+  BooleanField,
+  DropdownField,
+  NumericField,
+  ConnectionTypeField,
+  ConnectionTypeDataField,
+  ConnectionTypeConfigMap,
+  ConnectionTypeConfigMapObj,
+  ConnectionTypeValueType,
+  Connection,
+  ConnectionTypeFormData,
+  AWSDataEntry,
+} from './connectionTypes';
+
+export {
+  isConnectionTypeDataFieldType,
+  isConnectionTypeDataField,
+  isConnectionType,
+  isConnection,
+  getConnectionTypeRef,
+  toConnectionTypeConfigMapObj,
+  toConnectionTypeConfigMap,
+  defaultValueToString,
+  fieldTypeToString,
+  fieldNameToEnvVar,
+  ENV_VAR_NAME_REGEX,
+  isValidEnvVar,
+  ModelServingCompatibleTypes,
+  URIConnectionTypeKeys,
+  OCIConnectionTypeKeys,
+  OCIAccessTypeKey,
+  S3ConnectionTypeKeys,
+  isModelServingEnvVar,
+  getModelServingConnectionTypeName,
+  isModelServingCompatible,
+  getModelServingCompatibility,
+  filterModelServingConnectionTypes,
+  getDefaultValues,
+  getMRConnectionValues,
+  withRequiredFields,
+  assembleConnectionSecret,
+  parseConnectionSecretValues,
+  getConnectionTypeDisplayName,
+  filterEnabledConnectionTypes,
+  findSectionFields,
+  convertObjectStorageSecretData,
+  getConnectionProtocolType,
+} from './connectionTypeUtils';
