+++ packages/kserve/src/modelLocationData.ts
@@ -2,9 +2,8 @@ import {
   getPVCNameFromURI,
   isPVCUri,
 } from '@odh-dashboard/internal/pages/modelServing/screens/projects/utils';
-import { MetadataAnnotation } from '@odh-dashboard/k8s-core';
+import { MetadataAnnotation, ModelServingCompatibleTypes } from '@odh-dashboard/k8s-core';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
-import { ModelServingCompatibleTypes } from '@odh-dashboard/internal/concepts/connectionTypes/utils';
 import { ModelLocationData, ModelLocationType } from '@odh-dashboard/model-serving/types/form-data';
 
 export const getModelLocationUri = (deployment: InferenceServiceKind): string | undefined => {
