+++ packages/model-serving/modelRegistry/PreWizardDeployModal.tsx
@@ -14,13 +14,12 @@ import {
 } from '@patternfly/react-core';
 import { Link } from 'react-router';
 import { typedObjectImage, ProjectObjectType } from '@odh-dashboard/ui-core';
-import type { ProjectKind } from '@odh-dashboard/k8s-core';
+import type { ProjectKind, Connection } from '@odh-dashboard/k8s-core';
+import { getConnectionTypeRef } from '@odh-dashboard/k8s-core';
 import { ProjectsContext } from '@odh-dashboard/internal/concepts/projects/ProjectsContext';
 import ProjectSelector from '@odh-dashboard/internal/pages/modelServing/screens/projects/InferenceServiceModal/ProjectSelector';
 import useServingConnections from '@odh-dashboard/internal/pages/projects/screens/detail/connections/useServingConnections';
-import { Connection } from '@odh-dashboard/internal/concepts/connectionTypes/types';
 import { useWatchConnectionTypes } from '@odh-dashboard/internal/utilities/useWatchConnectionTypes';
-import { getConnectionTypeRef } from '@odh-dashboard/internal/concepts/connectionTypes/utils';
 import { uriToModelLocation } from '@odh-dashboard/internal/concepts/modelRegistry/utils';
 import { modelVersionDeploymentsUrl } from '@odh-dashboard/model-registry/routeUtils';
 import { ModelDeployPrefillInfo } from '@odh-dashboard/model-serving/shared';
