+++ packages/model-serving/modelRegistry/useExtractFormDataFromRegistry.ts
@@ -6,12 +6,11 @@ import {
   getResourceNameFromK8sResource,
   INFERENCE_SERVICE_NAME_INVALID_CHARS_MESSAGE,
   INFERENCE_SERVICE_NAME_REGEX,
+  AccessTypes,
 } from '@odh-dashboard/k8s-core';
+import type { ConnectionTypeValueType } from '@odh-dashboard/k8s-core';
 import { uriToModelLocation } from '@odh-dashboard/internal/concepts/modelRegistry/utils';
 import { useWatchConnectionTypes } from '@odh-dashboard/internal/utilities/useWatchConnectionTypes';
-import { ConnectionTypeValueType } from '@odh-dashboard/internal/concepts/connectionTypes/types';
-// eslint-disable-next-line @odh-dashboard/no-restricted-imports
-import { AccessTypes } from '@odh-dashboard/internal/pages/projects/dataConnections/const';
 import { ModelDeployPrefillInfo } from '@odh-dashboard/model-serving/shared';
 import { getModelRegistryMetadata } from './utils/deployUtils';
 import {
