+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/CustomTypeSelectField.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Flex, FlexItem, FormGroup, HelperTextItem, FormHelperText } from '@patternfly/react-core';
-import { ConnectionTypeConfigMapObj } from '@odh-dashboard/internal/concepts/connectionTypes/types';
+import type { ConnectionTypeConfigMapObj } from '@odh-dashboard/k8s-core';
 import SimpleSelect from '@odh-dashboard/ui-core/components/SimpleSelect';
 
 type CustomTypeSelectFieldProps = {
