+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/URIConnectionField.tsx
@@ -7,10 +7,7 @@ import {
   ValidatedOptions,
 } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
-import {
-  ConnectionTypeDataField,
-  UriField,
-} from '@odh-dashboard/internal/concepts/connectionTypes/types';
+import type { ConnectionTypeDataField, UriField } from '@odh-dashboard/k8s-core';
 import DefaultValueTextRenderer from '@odh-dashboard/internal/concepts/connectionTypes/fields/DefaultValueTextRenderer';
 import { trimInputOnBlur, trimInputOnPaste } from '@odh-dashboard/ui-core/utilities';
 import { FieldMode } from '@odh-dashboard/internal/concepts/connectionTypes/fields/types';
