+++ packages/model-serving/src/components/deploymentWizard/types.ts
@@ -3,12 +3,14 @@ import type { z } from 'zod';
 import type { K8sResourceCommon } from '@openshift/dynamic-plugin-sdk-utils';
 import type { useHardwareProfileConfig } from '@odh-dashboard/hardware-profiles/shared';
 import type { useK8sNameDescriptionFieldData } from '@odh-dashboard/ui-core/components/K8sNameDescriptionField';
-import {
+import type {
   ConnectionTypeConfigMapObj,
   ConnectionTypeValueType,
-} from '@odh-dashboard/internal/concepts/connectionTypes/types';
-import type { RecursivePartial } from '@odh-dashboard/foundation';
-import type { ProjectKind, SecretKind, SupportedModelFormats } from '@odh-dashboard/k8s-core';
+  ProjectKind,
+  RecursivePartial,
+  SecretKind,
+  SupportedModelFormats,
+} from '@odh-dashboard/k8s-core';
 import type { LabeledConnection } from '@odh-dashboard/internal/pages/modelServing/screens/types';
 import type { SimpleSelectOption } from '@odh-dashboard/ui-core/components/SimpleSelect';
 import type {
