+++ packages/model-serving/src/components/deploymentWizard/utils.ts
@@ -3,18 +3,12 @@ import {
   getGeneratedSecretName,
   getDisplayNameFromK8sResource,
   getResourceNameFromK8sResource,
-} from '@odh-dashboard/k8s-core';
-import type { SecretKind } from '@odh-dashboard/k8s-core';
-import {
   getConnectionTypeRef,
   getModelServingCompatibility,
   getModelServingConnectionTypeName,
   ModelServingCompatibleTypes,
-} from '@odh-dashboard/internal/concepts/connectionTypes/utils';
-import {
-  Connection,
-  ConnectionTypeConfigMapObj,
-} from '@odh-dashboard/internal/concepts/connectionTypes/types';
+} from '@odh-dashboard/k8s-core';
+import type { SecretKind, Connection, ConnectionTypeConfigMapObj } from '@odh-dashboard/k8s-core';
 import { type TokenAuthenticationFieldData } from './fields/TokenAuthenticationField';
 import {
   ModelLocationType,
